# Wildlife-Tracker-replit-1/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wildlife Track API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  /** BR or US */
  nationality: string;
  phone: string;
  vehiclePlate: string;
  /** cpf, rg, or driver_license */
  documentType: string;
  documentNumber: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  nationality: string;
  phone: string;
  vehiclePlate: string;
  documentType: string;
  documentNumber: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CreateSightingBody {
  animalType: string;
  /** @nullable */
  description?: string | null;
  latitude: number;
  longitude: number;
}

export interface Sighting {
  id: number;
  userId: number;
  userName: string;
  animalType: string;
  /** @nullable */
  description?: string | null;
  latitude: number;
  longitude: number;
  crowdCount: number;
  createdAt: string;
}

export interface SightingStatItem {
  animalType: string;
  count: number;
  lastSeen: string;
}

export interface SightingStats {
  total: number;
  byAnimalType: SightingStatItem[];
  activeEmergencies: number;
}

export interface CrowdStatus {
  sightingId: number;
  crowdCount: number;
  threshold: number;
  suppressed: boolean;
}

export interface CreateSubscriptionBody {
  animalType: string;
}

export interface Subscription {
  id: number;
  userId: number;
  animalType: string;
  createdAt: string;
}

export interface Notification {
  id: number;
  userId: number;
  /** sighting or emergency */
  type: string;
  message: string;
  /** @nullable */
  sightingId?: number | null;
  /** @nullable */
  emergencyId?: number | null;
  read: boolean;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export interface CreateEmergencyBody {
  description: string;
  latitude: number;
  longitude: number;
}

export interface Emergency {
  id: number;
  userId: number;
  userName: string;
  description: string;
  latitude: number;
  longitude: number;
  resolved: boolean;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export type ListSightingsParams = {
  /**
   * Filter by animal type
   */
  animalType?: string;
  /**
   * Max results
   */
  limit?: number;
};

export type GetRecentSightingsParams = {
  limit?: number;
};