# Wildlife-Tracker-replit-1/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wildlife Track API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: sightings
    description: Wildlife sightings
  - name: subscriptions
    description: Animal type subscriptions
  - name: notifications
    description: User notifications
  - name: emergencies
    description: Emergency alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sightings:
    get:
      operationId: listSightings
      tags: [sightings]
      summary: List all sightings
      parameters:
        - in: query
          name: animalType
          schema:
            type: string
          description: Filter by animal type
        - in: query
          name: limit
          schema:
            type: integer
          description: Max results
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sighting"
    post:
      operationId: createSighting
      tags: [sightings]
      summary: Report a new sighting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSightingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sighting"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sightings/stats:
    get:
      operationId: getSightingStats
      tags: [sightings]
      summary: Get sighting statistics grouped by animal type
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SightingStats"

  /sightings/recent:
    get:
      operationId: getRecentSightings
      tags: [sightings]
      summary: Get the most recent sightings
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sighting"

  /sightings/{id}:
    get:
      operationId: getSighting
      tags: [sightings]
      summary: Get a sighting by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sighting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sightings/{id}/crowd:
    get:
      operationId: getSightingCrowd
      tags: [sightings]
      summary: Get crowd status for a sighting location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrowdStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: Get current user subscriptions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe to an animal type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Already subscribed or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscriptions/{id}:
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Unsubscribe from an animal type
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user notifications
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      tags: [notifications]
      summary: Get count of unread notifications
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"

  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emergencies:
    get:
      operationId: listEmergencies
      tags: [emergencies]
      summary: List active emergencies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Emergency"
    post:
      operationId: createEmergency
      tags: [emergencies]
      summary: Trigger an emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmergencyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emergencies/{id}/resolve:
    patch:
      operationId: resolveEmergency
      tags: [emergencies]
      summary: Resolve an emergency alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        nationality:
          type: string
          description: "BR or US"
        phone:
          type: string
        vehiclePlate:
          type: string
        documentType:
          type: string
          description: "cpf, rg, or driver_license"
        documentNumber:
          type: string
      required:
        - email
        - password
        - name
        - nationality
        - phone
        - vehiclePlate
        - documentType
        - documentNumber

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        nationality:
          type: string
        phone:
          type: string
        vehiclePlate:
          type: string
        documentType:
          type: string
        documentNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - nationality
        - phone
        - vehiclePlate
        - documentType
        - documentNumber
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    CreateSightingBody:
      type: object
      properties:
        animalType:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
      required:
        - animalType
        - latitude
        - longitude

    Sighting:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        animalType:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
        crowdCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - animalType
        - latitude
        - longitude
        - crowdCount
        - createdAt

    SightingStatItem:
      type: object
      properties:
        animalType:
          type: string
        count:
          type: integer
        lastSeen:
          type: string
      required:
        - animalType
        - count
        - lastSeen

    SightingStats:
      type: object
      properties:
        total:
          type: integer
        byAnimalType:
          type: array
          items:
            $ref: "#/components/schemas/SightingStatItem"
        activeEmergencies:
          type: integer
      required:
        - total
        - byAnimalType
        - activeEmergencies

    CrowdStatus:
      type: object
      properties:
        sightingId:
          type: integer
        crowdCount:
          type: integer
        threshold:
          type: integer
        suppressed:
          type: boolean
      required:
        - sightingId
        - crowdCount
        - threshold
        - suppressed

    CreateSubscriptionBody:
      type: object
      properties:
        animalType:
          type: string
      required:
        - animalType

    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        animalType:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - animalType
        - createdAt

    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          description: "sighting or emergency"
        message:
          type: string
        sightingId:
          type: ["integer", "null"]
        emergencyId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - message
        - read
        - createdAt

    UnreadCount:
      type: object
      properties:
        count:
          type: integer
      required:
        - count

    CreateEmergencyBody:
      type: object
      properties:
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - description
        - latitude
        - longitude

    Emergency:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        resolved:
          type: boolean
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - userName
        - description
        - latitude
        - longitude
        - resolved
        - createdAt